OP-21546: Implementation OPA validation for static policies
